refactor: Update styling for pricing card options
This commit updates the styling for the pricing card options in the _pricing-card.scss file. The justify-content property is changed from center to flex-start, aligning the options to the left side of the card. This change improves the visual layout of the pricing cards and enhances the overall design of the website.