feat: Add favicon to the website
This commit adds a favicon to the website by including a link to the favicon file in the head section of the layout.jsx file. The favicon file is located at "/assets/img/logo.svg". This enhancement improves the visual appearance of the website and provides a recognizable icon for users when bookmarking or saving the website.